Split template generation from config writing
